+++ src/main/java/name/divinityunbound/block/ModBlocks.java
@@ -247,6 +247,9 @@ public class ModBlocks {
     public static final Block DEMETERS_HARVESTER = registerBlock("demeters_harvester",
             new DemetersHarvesterBlock(FabricBlockSettings.copyOf(Blocks.IRON_BLOCK).nonOpaque()));
 
+    public static final Block PERSEPHONES_BLESSING = registerBlock("persephones_blessing",
+            new PersephonesBlessingBlock(FabricBlockSettings.copyOf(Blocks.IRON_BLOCK).nonOpaque()));
+
     public static final Block ITEM_SINGULARITY_STORAGE = registerBlock("item_singularity_storage",
             new ItemSingularityStorageBlock(FabricBlockSettings.copyOf(Blocks.IRON_BLOCK).nonOpaque()));