+++ src/main/java/name/divinityunbound/block/entity/DemetersHarvesterBlockEntity.java
@@ -1,20 +1,27 @@
 package name.divinityunbound.block.entity;
 
-import name.divinityunbound.screen.ZeusBatteryScreenHandler;
+import name.divinityunbound.screen.DemetersHarvesterScreenHandler;
 import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
+import net.fabricmc.fabric.api.transfer.v1.item.InventoryStorage;
 import net.fabricmc.fabric.api.transfer.v1.transaction.Transaction;
+import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.CropBlock;
-import net.minecraft.block.TorchflowerBlock;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.SimpleInventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.ItemUsageContext;
+import net.minecraft.item.Items;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtList;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.network.listener.ClientPlayPacketListener;
 import net.minecraft.network.packet.Packet;
 import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
+import net.minecraft.registry.tag.ItemTags;
 import net.minecraft.screen.PropertyDelegate;
 import net.minecraft.screen.ScreenHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -23,14 +30,14 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.world.ModifiableWorld;
 import net.minecraft.world.World;
+import net.minecraft.world.event.GameEvent;
 import org.jetbrains.annotations.Nullable;
 import team.reborn.energy.api.base.SimpleEnergyStorage;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class DemetersHarvesterBlockEntity extends BlockEntity implements ExtendedScreenHandlerFactory {
-    private static final int DEFAULT_RANGE = 12;
+    private static final int DEFAULT_RANGE = 8;
 
     protected final PropertyDelegate propertyDelegate;
     private int progress = 0;
@@ -65,7 +72,30 @@ public class DemetersHarvesterBlockEntity extends BlockEntity implements Extende
         };
     }
 
-    public final SimpleEnergyStorage energyStorage = new SimpleEnergyStorage(5000000, Integer.MAX_VALUE, Integer.MAX_VALUE) {
+    public final SimpleInventory internalHoeInventory = new SimpleInventory(1) {
+        @Override
+        public boolean isValid(int slot, ItemStack stack) {
+            if (slot == 0) {
+                return stack.isIn(ItemTags.HOES);
+            }
+            return false;
+        }
+
+        @Override
+        public void markDirty() {
+            DemetersHarvesterBlockEntity.this.markDirty();
+        }
+    };
+    public final SimpleInventory internalInventory = new SimpleInventory(9) {
+        @Override
+        public void markDirty() {
+            DemetersHarvesterBlockEntity.this.markDirty();
+        }
+    };
+    public final InventoryStorage inventoryHoeWrapper = InventoryStorage.of(internalHoeInventory, null);
+    public final InventoryStorage inventoryWrapper = InventoryStorage.of(internalInventory, null);
+
+    public final SimpleEnergyStorage energyStorage = new SimpleEnergyStorage(500000, Integer.MAX_VALUE, Integer.MAX_VALUE) {
         @Override
         protected void onFinalCommit() {
             markDirty();
@@ -76,6 +106,8 @@ public class DemetersHarvesterBlockEntity extends BlockEntity implements Extende
     @Override
     protected void writeNbt(NbtCompound nbt) {
         super.writeNbt(nbt);
+        nbt.put("demeters_harvester.internalHoeInventory", internalHoeInventory.toNbtList());
+        nbt.put("demeters_harvester.internalInventory", internalInventory.toNbtList());
         nbt.putInt("demeters_harvester.progress", progress);
         nbt.putInt("demeters_harvester.speedCount", speedCount);
         nbt.putInt("demeters_harvester.quantityCount", quantityCount);
@@ -86,6 +118,8 @@ public class DemetersHarvesterBlockEntity extends BlockEntity implements Extende
     @Override
     public void readNbt(NbtCompound nbt) {
         super.readNbt(nbt);
+        internalHoeInventory.readNbtList((NbtList) nbt.get("demeters_harvester.internalHoeInventory"));
+        internalInventory.readNbtList((NbtList) nbt.get("demeters_harvester.internalInventory"));
         progress = nbt.getInt("demeters_harvester.progress");
         speedCount = nbt.getInt("demeters_harvester.speedCount");
         quantityCount = nbt.getInt("demeters_harvester.quantityCount");
@@ -113,7 +147,7 @@ public class DemetersHarvesterBlockEntity extends BlockEntity implements Extende
     @Nullable
     @Override
     public ScreenHandler createMenu(int syncId, PlayerInventory playerInventory, PlayerEntity player) {
-        return null;//new DemetersHarvesterScreenHandler(syncId, playerInventory, this, this.propertyDelegate);
+        return new DemetersHarvesterScreenHandler(syncId, playerInventory, this, this.propertyDelegate);
     }
 
     private int cooldown = 0;
@@ -121,61 +155,80 @@ public class DemetersHarvesterBlockEntity extends BlockEntity implements Extende
         if (world.isClient()) {
             return;
         }
-        if (cooldown < 20) {
+        if (cooldown < (60 - (speedCount * 6))) {
             cooldown++;
             return;
         }
         cooldown = 0;
 
-        int range = (DEFAULT_RANGE + (rangeCount*2)) / 2;
-        int posx = pos.getX();
-        int posy = pos.getY();
-        int posz = pos.getZ();
-        Box box = new Box(posx - range, posy - range, posz - range,
-                posx + range, posy + range, posz + range);
+        int range = (DEFAULT_RANGE + (rangeCount * 2)) / 2;
+        if (hasItemInHoeSlot()) {
+            harvestCrops(world, pos, range);
+            plantCrops(world, pos, range);
+        }
+    }
 
-        List<BlockPos> blockPositions = BlockPos.stream(-4, 0, -4, 4, 0, 4).filter((position) -> {
-            return Math.abs(position.getX()) == 4 || Math.abs(position.getZ()) == 4;
-        }).map(BlockPos::toImmutable).toList();
-
-        for (int i = -4; i <= 4; i++) {
-            for (int j = -4; j <= 4; j++) {
+    private void harvestCrops(World world, BlockPos pos, int range) {
+        for (int i = -range; i <= range; i++) {
+            for (int j = -range; j <= range; j++) {
                 BlockPos offsetPos = pos.add(i, 0, j);
                 BlockState blockState = world.getBlockState(offsetPos);
                 if (blockState.getBlock() instanceof CropBlock) {
                     if (((CropBlock) blockState.getBlock()).getAge(blockState) ==
                             ((CropBlock) blockState.getBlock()).getMaxAge()) {
                         ((ModifiableWorld) world).breakBlock(offsetPos, true);
-                    }
-                    else if (blockState.getBlock() instanceof TorchflowerBlock) {
-                        if (((CropBlock) blockState.getBlock()).getAge(blockState) ==
-                                ((CropBlock) blockState.getBlock()).getMaxAge() - 1) {
-                            ((ModifiableWorld) world).breakBlock(offsetPos, true);
-                        }
+                        damageHoe();
                     }
                 }
                 else if (blockState.getBlock().equals(Blocks.MELON) ||
                         blockState.getBlock().equals(Blocks.PUMPKIN)) {
                     ((ModifiableWorld) world).breakBlock(offsetPos, true);
+                    damageHoe();
                 }
             }
         }
-//        blockPositions.forEach((offset) -> {
-//
-//        });
+    }
 
-//        List<BlockState> blocks = world.getStatesInBoxIfLoaded(box)
-//                .filter((blockState) -> blockState.getBlock() instanceof CropBlock)
-//                .toList();
-//
-//        if (!blocks.isEmpty()) {
-//            blocks.forEach((block) -> {
-//                if (((CropBlock) block.getBlock()).getAge(block) ==
-//                        ((CropBlock) block.getBlock()).getMaxAge()) {
-//                    //((ModifiableWorld) world).breakBlock(block)
-//                }
-//            });
-//        }
+    private void damageHoe() {
+        if(internalHoeInventory.getStack(0).damage(1, world.getRandom(), null)) {
+            internalHoeInventory.getStack(0).setCount(0);
+        }
+    }
+
+    private void plantCrops(World world, BlockPos pos, int range) {
+        for (int i = -range; i <= range; i++) {
+            for (int j = -range; j <= range; j++) {
+                BlockPos offsetPos = pos.add(i, -1, j);
+                BlockState blockState = world.getBlockState(offsetPos);
+                if (blockState.getBlock().equals(Blocks.FARMLAND)) {
+                    BlockState blockStateAbove = world.getBlockState(offsetPos.up());
+                    if (blockStateAbove.isAir()) {
+                        ItemStack seeds = getSeeds();
+                        if (!seeds.isEmpty()) {
+                            BlockState defaultBlockState = Block.getBlockFromItem(seeds.getItem()).getDefaultState();
+                            world.setBlockState(offsetPos.up(), defaultBlockState, Block.NOTIFY_ALL);
+                            world.emitGameEvent(GameEvent.BLOCK_PLACE, offsetPos.up(), GameEvent.Emitter.of(defaultBlockState));
+                            seeds.decrement(1);
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private ItemStack getSeeds() {
+        ItemStack itemStack = ItemStack.EMPTY;
+        for (int i = 0; i < internalInventory.size(); i++) {
+            itemStack = internalInventory.getStack(i);
+            if (!itemStack.isEmpty() && Block.getBlockFromItem(itemStack.getItem()) instanceof CropBlock) {
+                return itemStack;
+            }
+        }
+        return ItemStack.EMPTY;
+    }
+
+    private boolean hasItemInHoeSlot() {
+        return !internalHoeInventory.getStack(0).isEmpty() && internalHoeInventory.getStack(0).isIn(ItemTags.HOES);
     }
 
     private void extractEnergy() {