+++ src/main/java/name/divinityunbound/datagen/ModBlockTagProvider.java
@@ -77,6 +77,7 @@ public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.SPACE_FORGED_BLOCK)
                 .add(ModBlocks.ZEUS_BATTERY)
                 .add(ModBlocks.DEMETERS_HARVESTER)
+                .add(ModBlocks.PERSEPHONES_BLESSING)
                 .add(ModBlocks.SPACE_TIME_FURNACE);
 
         getOrCreateTagBuilder(BlockTags.NEEDS_STONE_TOOL)
@@ -116,6 +117,7 @@ public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.SPACE_FORGED_BLOCK)
                 .add(ModBlocks.ZEUS_BATTERY)
                 .add(ModBlocks.DEMETERS_HARVESTER)
+                .add(ModBlocks.PERSEPHONES_BLESSING)
                 .add(ModBlocks.ITEM_SINGULARITY_STORAGE);
 
         getOrCreateTagBuilder(BlockTags.NEEDS_DIAMOND_TOOL)